uint to int

+++ day_24.py
@@ -5,15 +5,15 @@ import scipy.ndimage as nd
 KERNEL = np.array([[0, 1, 0], [1, 0, 1], [0, 1, 0]])
 
 with open('input24', 'r') as data:
-    data = [[char == '#' for char in line] for line in data.read().splitlines()]
+    data = [[int(char == '#') for char in line] for line in data.read().splitlines()]
 
 bugs = np.array(data)
 states = set((bugs.tostring(), ))
 while True:
-    neighbor_count = nd.convolve(bugs.astype(np.uint8), KERNEL, mode="constant")
+    neighbor_count = nd.convolve(bugs, KERNEL, mode="constant")
     still_alive = bugs & (neighbor_count == 1)
     new_borns = ~bugs & ((neighbor_count == 2) | (neighbor_count == 1))
-    bugs = still_alive + new_borns
+    bugs = (still_alive + new_borns).astype(int)
     if (as_string := bugs.tostring()) in states:
         break
     states.add(as_string)
@@ -25,7 +25,7 @@ levels[-1]; levels[1] # First outer and inner level
 for _ in range(200):
     new = {}
     for level, bugs in tuple(levels.items()):
-        neighbor_count = nd.convolve(bugs.astype(np.uint8), KERNEL, mode="constant")
+        neighbor_count = nd.convolve(bugs, KERNEL, mode="constant")
         for outer, inner in ((0, (1, 2)), ((..., 0), (2, 1)),
                              (4, (3, 2)), ((..., 4), (2, 3))):
             neighbor_count[outer] += levels[level - 1][inner]
@@ -33,7 +33,7 @@ for _ in range(200):
 
         still_alive = bugs & (neighbor_count == 1)
         new_borns = ~bugs & ((neighbor_count == 2) | (neighbor_count == 1))
-        new[level] = still_alive + new_borns
+        new[level] = (still_alive + new_borns).astype(int)
         new[level][2, 2] = 0 # Center stays empty
     levels.update(new)
